git-svn-id: svn://svn.code.sf.net/p/jaer/code@3387 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ branches/pseye/host/java/src/uk/ac/imperial/pseye/PSEyeCameraFileInputStream.java
@@ -4,7 +4,6 @@
  */
 package uk.ac.imperial.pseye;
 
-import cl.eye.*;
 import java.io.EOFException;
 import java.io.File;
 import java.io.IOException;
@@ -48,7 +47,7 @@ public class PSEyeCameraFileInputStream extends AEFileInputStream {
         }
         int[] addr = packet.getAddresses();
         int[] ts = packet.getTimestamps();
-        int oldPosition = position();
+        int oldPosition = (int) position();
         EventRaw ev;
         int count = 0;
         for (int frame = 0; frame < nframes; frame++) {
